Feat: The stars remember their place and change
More detailed, the stars remember their last ubication, but with this,
these ones only increase when the new points are higher than the stars
per lvl.

For example, when a lvl has 2 stars obtained, and then you win it again
with 1, it will not change. But if you win with 3 stars, it will change
to 3.